Geometry class (S4) and methods

A geom stores the vertices that outline shape, and all additional information that characterise a geometry. A geom can be spatial, but does not have to be. A geom can either have absolute or relative values, where relative values specify the vertex position relative to the window slot.

Details

A geom either has the feature type

"point", when none of the vertices are connected to other vertices,

"line", where vertices with the same fid are connected following the sequence of their order, without the line closing in itself and

"polygon" according to the same definition, except that the line closes in itself.

Moreover, a geom does NOT have the slot extent, which characterises the minimum and maximum value of the vertex coordinates and which is thus derived "on the fly" from the vertices. Instead it has a reference window, which is sort of a second extent that may be bigger than extent and which determines the relative position of the vertices when plotting.

Slots

type

[character(1)]
the type of feature, recently either "point", "line" or "polygon".

coords

[data.frame(1)]
the fid, x and y coordinates per vertex.

attr

[data.frame(1)]
arbitrary number of attributes per unique id in coords

window

[data.frame(1)]
the minimum and maximum value in x and y dimension of the reference window in which the geom dwells.

scale

[character(1)]
whether the vertex coordinates are stored as "absolute" values, or "relative" to window.

crs

[character(1)]
the coordinate reference system in proj4 notation.

history

[list(.)]
a list of steps taken to derive the geom in focus.