Get the table of point attributes

Get tabular information of the attributes of points (incl. coordinates).

# S4 method for ANY
getPoints(x, ...)

# S4 method for geom
getPoints(x, ...)

# S4 method for Spatial
getPoints(x)

# S4 method for sf
getPoints(x)

# S4 method for ppp
getPoints(x)

# S4 method for Raster
getPoints(x)

# S4 method for matrix
getPoints(x)

Arguments

x
the object from which to derive the attribute table.

...
subset based on logical predicates defined in terms of the columns in x or a vector of booleans. Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept.

Value

A table of the point attributes of x or an object where the point table has been subsetted.

See also

Examples

getPoints(x = gtGeoms$polygon)

#> # A tibble: 15 x 3
#>        x     y   fid
#>    <dbl> <dbl> <int>
#>  1     0     0     1
#>  2    10     0     1
#>  3    10    10     1
#>  4     0    10     1
#>  5     0     0     1
#>  6     1     1     1
#>  7     1     2     1
#>  8     2     2     1
#>  9     2     1     1
#> 10     1     1     1
#> 11    11    11     2
#> 12    11    12     2
#> 13    12    12     2
#> 14    12    11     2
#> 15    11    11     2

getPoints(x = gtGeoms$point)

#> # A tibble: 12 x 3
#>        x     y   fid
#>    <dbl> <dbl> <int>
#>  1     0     0     1
#>  2     0    10     2
#>  3    10    10     3
#>  4    10     0     4
#>  5     1     1     5
#>  6     1     2     6
#>  7     2     2     7
#>  8     2     1     8
#>  9    11    11     9
#> 10    11    12    10
#> 11    12    12    11
#> 12    12    11    12

# for a raster object, the @point slot is extracted from its' compact storage
gtGeoms$grid$continuous@point

#> # A tibble: 3 x 2
#>       x     y
#>   <dbl> <dbl>
#> 1     0     0
#> 2    60    56
#> 3     1     1

getPoints(x = gtGeoms$grid$continuous)

#> # A tibble: 3,360 x 3
#>      fid     x     y
#>    <int> <dbl> <dbl>
#>  1     1   0.5   0.5
#>  2     2   1.5   0.5
#>  3     3   2.5   0.5
#>  4     4   3.5   0.5
#>  5     5   4.5   0.5
#>  6     6   5.5   0.5
#>  7     7   6.5   0.5
#>  8     8   7.5   0.5
#>  9     9   8.5   0.5
#> 10    10   9.5   0.5
#> # … with 3,350 more rows