Get the spatial resolution of a spatial object.

Get the spatial resolution of a spatial object.

# S4 method for ANY
getRes(x)

# S4 method for geom
getRes(x)

# S4 method for Raster
getRes(x)

# S4 method for matrix
getRes(x)

Arguments

x
the object from which to derive the resolution.

Value

A vector of two values of the spatial resolution of x in x and y dimensio n.

See also

Examples

getRes(x = gtGeoms$grid$continuous)

#> [1] 1 1

getRes(x = gtRasters$categorical)

#> [1] 1 1

getRes(x = matrix(0, 3, 5))

#> [1] 1 1