Get the subset of a spatial object.

Get the subset of a spatial object.

# S4 method for geom
getSubset(x, ..., slot = "table")

# S4 method for Spatial
getSubset(x, ...)

# S4 method for sf
getSubset(x, ...)

Arguments

x
object to subset.

...
Logical predicates defined in terms of the variables in x. Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept.

slot
[character(1)]
the slot in which to determine a subset, either "table" or "coords".

Examples

(obj <- gtGeoms$point)

#> class      : geom
#> type       : point
#> features   : 8  (8 vertices)
#> window     : 0, 12, 0, 12  (xmin, xmax, ymin, ymax)
#> extent     : 0, 12, 0, 12  (xmin, xmax, ymin, ymax)
#> scale      : absolute
#> crs        : NA
#> attributes : 2  (fid, n)

# get subset of features that have less than 3 vertices
(getSubset(x = gtGeoms$line, n < 5))

#> class      : geom
#> type       : line
#> features   : 1  (4 vertices)
#> window     : 0, 12, 0, 12  (xmin, xmax, ymin, ymax)
#> extent     : 1, 2, 1, 2  (xmin, xmax, ymin, ymax)
#> scale      : absolute
#> crs        : NA
#> attributes : 2  (fid, n)

(obj <- gtSP$SpatialPolygonsDataFrame)

#> class       : SpatialPolygonsDataFrame 
#> features    : 2 
#> extent      : 0, 12, 0, 12  (xmin, xmax, ymin, ymax)
#> coord. ref. : NA 
#> variables   : 1
#> names       : a 
#> min values  : 1 
#> max values  : 2 

(getSubset(x = obj, a == 2))

#> class       : SpatialPolygonsDataFrame 
#> features    : 1 
#> extent      : 11, 12, 11, 12  (xmin, xmax, ymin, ymax)
#> coord. ref. : NA 
#> variables   : 1
#> names       : a 
#> value       : 2 

(obj <- gtSF$polygon)

#> Simple feature collection with 2 features and 1 field
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a                       geometry
#> 1 1 POLYGON ((0 0, 10 0, 10 10,...
#> 2 2 POLYGON ((11 11, 11 12, 12 ...

(getSubset(x = obj, a == 1))

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 10 ymax: 10
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a                       geometry
#> 1 1 POLYGON ((0 0, 10 0, 10 10,...