Get the attribute table of a spatial object.

Get the attribute table of a spatial object.

# S4 method for geom
getTable(x, slot = "feat")

# S4 method for Spatial
getTable(x)

# S4 method for sf
getTable(x)

# S4 method for RasterLayer
getTable(x)

Arguments

x
the object from which to derive the attribute table.

slot
[character(1)]
the geom slot from which to retrieve the attribute table, either "vert", "feat" or "group".

Examples

getTable(gtGeoms$polygon)

#> # A tibble: 2 x 2
#>     fid   gid
#>   <int> <int>
#> 1     1     1
#> 2     2     2

getTable(x = gtSP$SpatialPolygons)

#> # A tibble: 2 x 2
#>     fid   gid
#>   <int> <int>
#> 1     1     1
#> 2     2     2

getTable(gtSF$multiline)

#> # A tibble: 3 x 3
#>     fid   gid     a
#>   <int> <int> <dbl>
#> 1     1     1     1
#> 2     2     1     1
#> 3     3     2     2

getTable(gtRasters$categorical)

#> # A tibble: 9 x 2
#>      id cover            
#>   <int> <fct>            
#> 1     1 lake             
#> 2    11 sealed           
#> 3    21 wetland          
#> 4    24 arable           
#> 5    27 pasture          
#> 6    31 clearcut         
#> 7    41 deciduous forest 
#> 8    44 coniferous forest
#> 9    47 mixed forest