Get the table of coordinates of a spatial object.

Get the table of coordinates of a spatial object.

# S4 method for geom
getVertices(x)

# S4 method for Spatial
getVertices(x)

# S4 method for sf
getVertices(x)

# S4 method for ppp
getVertices(x)

Arguments

x
the object from which to extract the coordinates

Value

A table of the coordinates x is made up of.

Examples

getVertices(gtGeoms$polygon)

#> # A tibble: 15 x 3
#>        x     y   fid
#>    <dbl> <dbl> <int>
#>  1     0     0     1
#>  2    10     0     1
#>  3    10    10     1
#>  4     0    10     1
#>  5     0     0     1
#>  6     1     1     1
#>  7     1     2     1
#>  8     2     2     1
#>  9     2     1     1
#> 10     1     1     1
#> 11    11    11     2
#> 12    11    12     2
#> 13    12    12     2
#> 14    12    11     2
#> 15    11    11     2

getVertices(gtSP$SpatialPoints)

#> # A tibble: 4 x 3
#>       x     y   fid
#>   <dbl> <dbl> <int>
#> 1     0     0     1
#> 2    10     0     2
#> 3    10    10     3
#> 4     0    10     4

getVertices(gtSF$multilinestring)

#> # A tibble: 12 x 3
#>        x     y   fid
#>    <dbl> <dbl> <int>
#>  1     0     0     1
#>  2     0    10     1
#>  3    10    10     1
#>  4    10     0     1
#>  5     1     1     2
#>  6     1     2     2
#>  7     2     2     2
#>  8     2     1     2
#>  9    11    11     3
#> 10    11    12     3
#> 11    12    12     3
#> 12    12    11     3

getVertices(gtPPP)

#> # A tibble: 15 x 3
#>        x     y   fid
#>    <dbl> <dbl> <int>
#>  1     0     0     1
#>  2    10     0     2
#>  3    10    10     3
#>  4     0    10     4
#>  5     0     0     5
#>  6     1     1     6
#>  7     1     2     7
#>  8     2     2     8
#>  9     2     1     9
#> 10     1     1    10
#> 11    11    11    11
#> 12    11    12    12
#> 13    12    12    13
#> 14    12    11    14
#> 15    11    11    15