Derive a geom from other spatial objects

See geom-class for details on differences between objects of class geom and other spatial classes.

gt_as_geom(input)

Arguments

input
[Spatial* | sf]
the spatial object to build a geom from.

Value

a geom of the type that comes closest to the type of the input.

Details

bla

Examples

spPoly <- rtSP$SpatialPolygons

#> Error in eval(expr, envir, enclos): object 'rtSP' not found

plot(spPoly, col = "goldenrod")

#> Error in plot(spPoly, col = "goldenrod"): object 'spPoly' not found

visualise(geom = gFrom(input = spPoly))

#> Error in gFrom(input = spPoly): could not find function "gFrom"

sfPoly <- rtSF$polygon

#> Error in eval(expr, envir, enclos): object 'rtSF' not found

plot(sfPoly, col = "goldenrod")

#> Error in plot(sfPoly, col = "goldenrod"): object 'sfPoly' not found

visualise(geom = gFrom(input = sfPoly))

#> Error in gFrom(input = sfPoly): could not find function "gFrom"