Derive bounding container of geometric objects

Enclose all vertices in different types of bounding containers.

gt_enclose(obj, fid = NULL, type = "convex")

Arguments

obj
[geometric object(1)]
the object for which to derive a bounding container.

fid
[integerish(.)]
in case a bounding container shall be derived only for a subset of features, specify that here.

type
[character(1)]
the type of bounding container to derive, by default this is the convex hull. Other options are rectangle, box, diamond, ball and ellipsoid, for explanations, see Details.

Value

geom of the bounding container.

Details

Bounding containers are geometric objects of the smallest area that encloses a set of vertices.

The convex hull is the smallest bounding container, as it encloses all vertices along the outer edges, and is contained in any other bounding container.

The minimal bounding rectangle is the enclosing rectangle with the smallest area and an arbitrary orientation.

The bounding box is a special case of the minimal bounding rectangle, which is derived from the minimum and maximum values in x and y dimension and is thus "parallel" to the x and y axis.

The bounding diamond is also a special case of the minimal bounding rectangle, where the lines have a slope of -1 or 1 (i.e., which is rotated by 45°).

The bounding ball, or minimal spanning sphere, is the smallest circle that encloses all vertices.

The bounding ellipsoid, or minimal spanning ellipsoid, is the smallest ellipsoid that encloses all vertices.

See also