Intersect geometric objects

Intersect geometric objects

gt_intersect(obj, fid = NULL, identify = TRUE)

Arguments

obj
[geometric object(1)]
the object to intersect.

fid
[integerish(.)]
in case only a subset of features shall be intersected, specify that here.

identify
[logical(1)]
whether or not to identify from which input features the intersected features originate. This is saved into the features attribute table and can be used to carry out set operations, see Details.

Value

geom of the intersected input objects.

Set-theoretic operations

At the basis of most set operations is an intersection of the involved sets, which allows to recombine the members into new sets. In geometr this is handled via gt_intersect(..., identify = TRUE), which attaches the parent membership of each member of the intersected set to the features attribute table, ready for subsetting.

For example, when two features are intersected, the intersection would have the "set"-attribute "1 2" (according to the feature ID of the parent members), while the non-overlapping parts would have the "set"-attribute "1" and "2" respectively. Moreover, the number of memberships is pasted into the "parent"-attribute, where the intersection would have the value 2 (because it originates from an intersection of 2 parent members) and the non-overlapping parts would have the number 1. This allows to carry out all basic and more complicated set operations via selecting the respective subsets with gt_filter and merging members into new sets via gt_dissolve.

union:

intersection:

difference:

symmetric difference:

See also