Rotate geometries

Rotate geoms by a certain angle about a center

gt_rotate(geom, angle, about = c(0, 0), fid = NULL)

Arguments

geom
[geom]
Object of class geom.

angle
[numeric(1)]
the counter-clockwise angle by which geom shall be rotated.

about
[numeric(2)]
the point about which geom shall be rotated.

fid
[integerish(.)]
vector of features that should be rotated.

Value

Rotated geom.

Examples

coords <- data.frame(x = c(30, 60, 60, 40, 10, 40, 20),
                     y = c(40, 40, 60, 70, 10, 20, 40),
                     fid = c(1, 1, 1, 1, 2, 2, 2))
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
aGeom <- geomPolygon(anchor = coords, window = window, show = TRUE)

#> Error in geomPolygon(anchor = coords, window = window, show = TRUE): could not find function "geomPolygon"

rotatedGeom <- gRotate(geom = aGeom, angle = 90, about = c(40, 40))

#> Error in gRotate(geom = aGeom, angle = 90, about = c(40, 40)): could not find function "gRotate"

visualise(geom = rotatedGeom)

#> Error in visualise(geom = rotatedGeom): object 'rotatedGeom' not found

# rotate single objects
rotatedTriangle <- gRotate(geom = aGeom, angle = -180, about = c(30, 40), fid = 2)

#> Error in gRotate(geom = aGeom, angle = -180, about = c(30, 40), fid = 2): could not find function "gRotate"

visualise(geom = rotatedTriangle, col = "goldenrod1")

#> Error in visualise(geom = rotatedTriangle, col = "goldenrod1"): object 'rotatedTriangle' not found

# rotate different objects about different centers by different angles
rotateMore <- gRotate(geom = aGeom,
                      angle = list(90, -180),
                      about = list(c(40, 40), c(30, 40)))

#> Error in gRotate(geom = aGeom, angle = list(90, -180), about = list(c(40,     40), c(30, 40))): could not find function "gRotate"

visualise(geom = rotateMore, col = "deeppink")

#> Error in visualise(geom = rotateMore, col = "deeppink"): object 'rotateMore' not found