Locate (and identify) clicks

Click into a plot to get the location or identify values

locate(samples = 1, raster = NULL, panel = NULL, identify = FALSE,
  snap = FALSE, raw = FALSE, silent = FALSE, show = FALSE, ...)

Arguments

samples
[integerish(1)]
the number of clicks.

raster
[Raster* | matrix]
the raster object in which to locate.

panel
[character(1)]
the panel in which to locate (i.e. the title shown over the plot). Does not have to be accurate, as it matches with grep.

identify
[logical(1)]
get the raster value at the sampled location (TRUE) or merely the location (FALSE, default).

snap
[logical(1)]
should the returned value(s) be set to the nearest grid cell's center (TRUE) or should they remain the selected, "real" value (FALSE, default)?

raw
[logical(1)]
should the complete statistics about the clicks be returned (TRUE), or should only the basic output be returned (FALSE, default)?

silent
[logical(1)]
should all messages except errors be suppressed (TRUE), or should all messages be printed (FALSE, default)?

show
[logical(1)]
should information about the selected cells be included in the plot (TRUE), or should they merely be returned in the console (FALSE, default)?

...
[various]
graphical parameters of the objects that are created when show = TRUE.

Value

a data.frame of the selected locations and, if identify = TRUE, the respective values. If show = TRUE the values are also shown in the plot.

Examples

# NOT RUN {
input <- rtRasters$continuous
patches <- rPatches(rBinarise(input, thresh = 30))

# locate from a not yet plotted object
locate(samples = 3, raster = input, show = TRUE, col = "green")

# or from one that is already plotted
visualise(raster = input)
locate(identify = TRUE, snap = TRUE)

# with several panels, specify a target
visualise(raster = raster::brick(input, patches))
locate(samples = 4, panel = "patches", identify = TRUE, show = TRUE)
# }