Set the attribute table(s) of a spatial object.

Set the attribute table(s) of a spatial object.

# S4 method for geom
setTable(x, table = NULL, slot = "feat")

# S4 method for Spatial
setTable(x, table = NULL)

# S4 method for sf
setTable(x, table = NULL)

# S4 method for sfc
setTable(x, table = NULL)

# S4 method for ppp
setTable(x, table = NULL)

# S4 method for RasterLayer
setTable(x, table = NULL)

Arguments

x
the object to which to assign an attribute table.

table
[data.frame(.)]
the attribute table.

slot
[character(1)]
the slot (of geom) for which to set the attribute table, either "vert", "feat" or "group".

Value

The object x with an updated attribute table.

Details

If table does not have columns in common with x, the new columns are simply bound to the original attribute table (if possible). If there are common columns, they are joined.

Examples

# set table of a geom
# individual attributes per point/line/polygon feature
getTable(gtGeoms$point)

#> # A tibble: 3 x 2
#>     fid   gid
#>   <dbl> <dbl>
#> 1     1     1
#> 2     2     2
#> 3     3     3

newAttr <- setTable(x = gtGeoms$point,
                    slot = "vert",
                    table = data.frame(attr = letters[c(1:12)]))
getTable(x = newAttr, slot = "vert")

#> # A tibble: 12 x 4
#>        x     y   fid attr 
#>    <dbl> <dbl> <dbl> <fct>
#>  1     0     0     1 a    
#>  2     0    10     1 b    
#>  3    10    10     1 c    
#>  4    10     0     1 d    
#>  5     1     1     2 e    
#>  6     1     2     2 f    
#>  7     2     2     2 g    
#>  8     2     1     2 h    
#>  9    11    11     3 i    
#> 10    11    12     3 j    
#> 11    12    12     3 k    
#> 12    12    11     3 l    

newAttr2 <- setTable(x = newAttr,
                     slot = "feat",
                     table = data.frame(gid = c(1:3), attr = letters[1:3]))

#> Joining, by = "gid"

getTable(x = newAttr2, slot = "feat")

#> # A tibble: 3 x 3
#>     fid   gid attr 
#>   <dbl> <dbl> <fct>
#> 1     1     1 a    
#> 2     2     2 b    
#> 3     3     3 c    

newAttr2

#> geom        point
#>             3 features | 12 vertices
#> crs         cartesian
#> attributes  (vertices) attr
#>             (features) attr
#> tiny map      12 
#>               ◯ ◌ ◌ ◉    
#>               ◌ ◌ ◌ ◌    
#>               ◌ ◌ ◌ ◌    
#>             0 ◉ ◌ ◌ ◯ 12 
#>               0 

# set table of an sf
sfObj <- gtSF$polygon

# ... with common columns
myAttributes <- data.frame(a = c(2, 1), attr = letters[1:2])
setTable(x = sfObj, table = myAttributes)

#> Joining, by = "a"

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a attr                       geometry
#> 1 1    b POLYGON ((0 0, 10 0, 10 10,...
#> 2 2    a POLYGON ((11 11, 11 12, 12 ...

# ... without common columns
setTable(x = sfObj, table = myAttributes[2])

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   attr a                       geometry
#> 1    a 1 POLYGON ((0 0, 10 0, 10 10,...
#> 2    b 2 POLYGON ((11 11, 11 12, 12 ...

# set table to an sfc (no join possible) and transform it to sf thereby
sfcObj = gtSF$polygon$geometry
setTable(x = sfcObj, table = myAttributes)

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a attr                           geom
#> 1 2    a POLYGON ((0 0, 10 0, 10 10,...
#> 2 1    b POLYGON ((11 11, 11 12, 12 ...

table <- data.frame(attr = LETTERS[1:15],
                    colour = topo.colors(15))
setTable(gtPPP, table = table)

#> Marked planar point pattern: 15 points
#> Mark variables: attr, colour 
#> window: rectangle = [0, 12] x [0, 12] units