Visualise raster and geom objects

Visualise raster and geom objects

visualise(raster = NULL, geom = NULL, theme = NULL, trace = FALSE,
  image = FALSE, new = TRUE, ...)

Arguments

raster
[Raster* | matrix]
raster object to plot.

geom
[geom]
geom to plot.

theme
[list(7)]
Visualising options; see setTheme for details.

trace
[logical(1)]
Print the raster object's history (i.e. the process according to which it has been created) (TRUE), or simply plot the object (FALSE, default).

image
[logical(1)]
Does raster have the channels red, green and blue, i.e. is it an "image" (TRUE) or is this not the case (FALSE, default)?

new
[logical(1)]
force a new plot (TRUE, default).

...
[various]
Graphical parameters to geom.

Value

Returns invisibly an object of class recordedplot, see recordPlot for details (and warnings).

Details

To create a plot with your own style, design it with setTheme and use it in theme.

In case you want to plot an image (simiar to plotRGB), you have to provide a RasterStack or RasterBrick with the three layers red, green and blue and set image = TRUE.

Examples

library(rasterTools)

#> Error: package ‘geometr’ required by ‘rasterTools’ could not be found

input <- rtRasters$continuous

#> Error in eval(expr, envir, enclos): object 'rtRasters' not found

binarised <- rBinarise(input, thresh = 40)

#> Error in rBinarise(input, thresh = 40): could not find function "rBinarise"

visualise(raster = rDistance(binarised), trace = TRUE)

#> Error in rDistance(binarised): could not find function "rDistance"

# visualise also RasterBrick/-Stack objects
getDistances <- list(disEuc = list(operator = "rDistance"),
                     disMht = list(operator = "rDistance",
                                   method = "manhattan"),
                     disChb = list(operator = "rDistance",
                                   method = "chessboard"))
distances <- modify(input = binarised, by = getDistances, merge = TRUE)

#> Error in modify(input = binarised, by = getDistances, merge = TRUE): could not find function "modify"

distances <- raster::brick(binarised, distances)

#> Error in raster::brick(binarised, distances): object 'binarised' not found

visualise(distances)

#> Error in visualise(distances): object 'distances' not found

# define a geometry
coords <- data.frame(x = c(30, 60, 60, 40),
                     y = c(40, 40, 60, 70),
                     fid = 1)
window <- data.frame(x = c(0, 80),
                     y = c(0, 80))
(aGeom <- gs_polygon(anchor = coords, window = window, col = "blue"))

#> class      : geom
#> type       : polygon
#> features   : 1  (4 vertices)
#> window     : 0, 80, 0, 80  (xmin, xmax, ymin, ymax)
#> extent     : 30, 60, 40, 70  (xmin, xmax, ymin, ymax)
#> scale      : absolute
#> crs        : NA
#> attributes : 2  (fid, n)

# if plotted on top of an existing plot, the relative coordinate values
# will be used to construct the grob.
visualise(raster = input, geom = aGeom)

#> Error in visualise(raster = input, geom = aGeom): object 'input' not found

visualise(geom = aGeom)

#> Error in isTRUE(lhs): trying to get slot "geom" from an object of a basic class ("NULL") with no slots