Get the subset of a spatial object.

Get the subset of a spatial object.

# S4 method for geom
getSubset(x, ..., slot = "feat")

# S4 method for Spatial
getSubset(x, ...)

# S4 method for sf
getSubset(x, ...)

# S4 method for ppp
getSubset(x)

Arguments

x
object to subset.

...
Logical predicates defined in terms of the variables in x or a vector of booleans. Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept.

slot
[character(1)]
the slot in which to determine a subset, either "vert" for vertices, "feat" for features or "group" for group tables.

Value

A subset of x in its original class.

See also

Examples

# get the subset of a geom
(obj <- gtGeoms$point)

#> geom        point
#>             3 features | 12 vertices
#> crs         cartesian
#> attributes  --
#> tiny map      12 
#>               ◯ ◌ ◌ ◉    
#>               ◌ ◌ ◌ ◌    
#>               ◌ ◌ ◌ ◌    
#>             0 ◉ ◌ ◌ ◯ 12 
#>               0 

# the first feature
(getSubset(x = gtGeoms$line, fid == 1))

#> geom        line
#>             1 feature | 4 vertices
#> crs         cartesian
#> attributes  --
#> tiny map      12 
#>               ◎ ◌ ◌ ◎    
#>               ◌ ◌ ◌ ◌    
#>               ◌ ◌ ◌ ◌    
#>             0 ◎ ◌ ◌ ◎ 12 
#>               0 

# get the subset of a Spatial object
(obj <- gtSP$SpatialPolygonsDataFrame)

#> class       : SpatialPolygonsDataFrame 
#> features    : 2 
#> extent      : 0, 12, 0, 12  (xmin, xmax, ymin, ymax)
#> crs         : NA 
#> variables   : 1
#> names       : a 
#> min values  : 1 
#> max values  : 2 

# the subset where the attribute a equals 2
(getSubset(x = obj, a == 2))

#> class       : SpatialPolygonsDataFrame 
#> features    : 1 
#> extent      : 11, 12, 11, 12  (xmin, xmax, ymin, ymax)
#> crs         : NA 
#> variables   : 1
#> names       : a 
#> value       : 2 

# get the subset of an sf object
(obj <- gtSF$polygon)

#> Simple feature collection with 2 features and 1 field
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a                       geometry
#> 1 1 POLYGON ((0 0, 10 0, 10 10,...
#> 2 2 POLYGON ((11 11, 11 12, 12 ...

# the subset where the attribute a equals 1
(getSubset(x = obj, a == 1))

#> Simple feature collection with 1 feature and 1 field
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 10 ymax: 10
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a                       geometry
#> 1 1 POLYGON ((0 0, 10 0, 10 10,...

# getSubset(gtPPP$...)