Subset a geometric object using column values

This function allows to subset any geometric object for which all required getters are available.

gt_filter(obj, ..., update = TRUE)

Arguments

obj
[geometric object(1)]
the object to derive a subset from.

...
subset based on logical predicates defined in terms of the columns in x (of both, points, features and groups). Multiple conditions are combined with &. Only rows where the condition evaluates to TRUE are kept.

update
[logical(1)]
whether or not to update the window slot after deriving the subset.

Value

geom of the subset of obj.

See also

Examples

# get a subset of a geom
gt_filter(gtGeoms$point, y < -10)

#> geom        point
#>             2 groups | 4 features | 4 points
#> crs         cartesian
#> attributes  --
#> tiny map           -20 
#>                    ◌ ◌ ◌ ◌       
#>                    ◌ ◌ ◌ ◌       
#>                    ◌ ◌ ◌ ◌       
#>             -34.64 ◌ ◎ ◎ ◌ 34.64 
#>                    -40 

# get a subset of an sf-object
gt_filter(obj = gtSF$multilinestring, a == 1)

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> old-style crs object detected; please recreate object with a recent sf::st_crs()

#> geom        line
#>             1 group | 2 features | 8 points
#> crs         cartesian
#> attributes  (features) a
#> tiny map      10 
#>               ○ ◌ ◌ ○    
#>               ◌ ◌ ◌ ◌    
#>               ◌ ◌ ◌ ◌    
#>             0 ◉ ◌ ◌ ○ 10 
#>               0