Set the attribute table of a spatial object.

Set the attribute table of a spatial object.

# S4 method for geom
setTable(x, table)

# S4 method for Spatial
setTable(x, table)

# S4 method for sf
setTable(x, table)

# S4 method for sfc
setTable(x, table)

# S4 method for RasterLayer
setTable(x, table)

Arguments

x
the object to which to assign table.

table
[data.frame(.)]
the attribute table.

Details

When setting the attribute table of a simple feature (sf) - which is already a "data.frame-like object" - this really means that a table is joined to the already existing table.

When setting the attribute table of a simple feature geometry column (sfc), this upgrades the object to an sf object.

Examples

# set table to an sf
sfObj <- gtSF$polygon

# ... with common columns
myAttributes <- data.frame(a = c(2, 1), b = c("X", "Y"))
setTable(x = sfObj, table = myAttributes)

#> Joining, by = "a"

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a b                       geometry
#> 1 1 Y POLYGON ((0 0, 10 0, 10 10,...
#> 2 2 X POLYGON ((11 11, 11 12, 12 ...

# ... without common columns
setTable(x = sfObj, table = myAttributes[2])

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   b a                       geometry
#> 1 X 1 POLYGON ((0 0, 10 0, 10 10,...
#> 2 Y 2 POLYGON ((11 11, 11 12, 12 ...

# set table to an sfc (no join possible)
sfcObj = gtSF$polygon$geometry
setTable(x = sfcObj, table = myAttributes)

#> Simple feature collection with 2 features and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 0 ymin: 0 xmax: 12 ymax: 12
#> epsg (SRID):    NA
#> proj4string:    NA
#>   a b                           geom
#> 1 2 X POLYGON ((0 0, 10 0, 10 10,...
#> 2 1 Y POLYGON ((11 11, 11 12, 12 ...